SmartHub/v2/requirements-connector.txt
paho-mqtt
confluent-kafka==1.7.0
kafka-python==2.0.2